add memory allocating with correct size based on rom size

+++ src/Rom.cpp
@@ -26,7 +26,7 @@ void Rom::LoadSize(std::ifstream &file)
     file.seekg(0, std::ios::beg);
 
     this->logger->Log("rom: size (mb)", std::to_string(
-        static_cast<double>(this->size) / (1024 * 1024)
+        this->size / (1024 * 1024)
     ));
 }
